+++ Cesium3DTilesSelection/src/Tile.cpp
@@ -983,10 +983,52 @@ void Tile::upsampleParent(
     void* pRendererResources;
   };
 
+  int32_t index = 0;
+
+  const gsl::span<Tile> children = pParent->getChildren();
+  if (std::get_if<QuadtreeTileID>(&pParent->getTileID()) &&
+      std::any_of(
+          children.begin(),
+          children.end(),
+          [](const Tile& tile) noexcept {
+            return std::get_if<QuadtreeTileID>(&tile.getTileID());
+          })) {
+    TileContentLoadResult& content = *pParent->getContent();
+    std::vector<Projection>& parentProjections =
+        content.overlayDetails->rasterOverlayProjections;
+    if (_pContext->implicitContext->projection.has_value()) {
+      const Projection& projection = *_pContext->implicitContext->projection;
+      auto it = std::find(
+          parentProjections.begin(),
+          parentProjections.end(),
+          projection);
+
+      if (it == parentProjections.end()) {
+        parentProjections.emplace_back(projection);
+        index = int32_t(parentProjections.size()) - 1;
+
+        const BoundingRegion* pParentRegion =
+            std::get_if<BoundingRegion>(&pParent->getBoundingVolume());
+
+        GltfContent::createRasterOverlayTextureCoordinates(
+            parentModel,
+            pParent->getTransform(),
+            index,
+            pParentRegion ? std::make_optional<GlobeRectangle>(
+                                pParentRegion->getRectangle())
+                          : std::nullopt,
+            {projection});
+      } else {
+        index = int32_t(it - parentProjections.begin());
+      }
+    }
+  }
+
   pTileset->getAsyncSystem()
       .runInWorkerThread(
           [&parentModel,
            transform = this->getTransform(),
+           textureCoordinateIndex = index,
            projections = std::move(projections),
            pSubdividedParentID,
            tileBoundingVolume = this->getBoundingVolume(),
@@ -1002,7 +1044,8 @@ void Tile::upsampleParent(
                 std::make_unique<TileContentLoadResult>();
             pContent->model = upsampleGltfForRasterOverlays(
                 parentModel,
-                *pSubdividedParentID);
+                *pSubdividedParentID,
+                textureCoordinateIndex);
 
             // We can't necessarily trust our original bounding volume, so
             // recompute it here. See: