+++ Cesium3DTilesSelection/src/upsampleGltfForRasterOverlays.h
@@ -7,6 +7,7 @@ namespace Cesium3DTilesSelection {
 
 CesiumGltf::Model upsampleGltfForRasterOverlays(
     const CesiumGltf::Model& parentModel,
-    CesiumGeometry::UpsampledQuadtreeNode childID);
+    CesiumGeometry::UpsampledQuadtreeNode childID,
+    int32_t textureCoordinateIndex = 0);
 
 }